% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint_sem}
\alias{confint_sem}
\title{Standard Error of the Mean}
\usage{
confint_sem(x, conf.level = 0.95)
}
\arguments{
\item{x}{a numeric vector or R object which is coercible to one}

\item{conf.level}{the confidence level (alpha) of the Interval}
}
\value{
a \code{data.frame} with the mean, SEM and its Confidence Interval
}
\description{
Standard Error of the Mean
}
\examples{
iq <- rnorm(100, 100, 15)

confint_sem(iq)
}

