% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_helpers.R
\name{drop_labels}
\alias{drop_labels}
\title{Re-label a vector after subsetting}
\usage{
drop_labels(x)
}
\arguments{
\item{x}{A vector with now unused labels}
}
\value{
Identical vector with appropriate labels
}
\description{
Re-label a vector after subsetting
}
\examples{
\dontrun{
x <- drop_labels(x)
}
}

