% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_stats.R
\name{etasq}
\alias{etasq}
\title{Just eta^2.}
\usage{
etasq(formula, data)
}
\arguments{
\item{formula}{The model formula for \code{\link[stats:aov]{stats::aov}}.}

\item{data}{The \code{data.frame}.}
}
\value{
A single numeric value
}
\description{
Thin wrapper for \code{\link[DescTools:EtaSq]{DescTools::EtaSq}} to retrieve $eta^2$ for a simple use case
without having to fit an \code{aov} model beforehand.
Only use this for a simple one-way design, e.g. only one independent variable.
}
\examples{
etasq(stunzahl ~ jahrgang, ngo)
}
