% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tadaa_pairwise_games_howell.R
\name{games_howell}
\alias{games_howell}
\title{Games Howell post-hoc procedure}
\source{
\url{https://gist.github.com/aschleg/ea7942efc6108aedfa9ec98aeb6c2096}
}
\usage{
games_howell(obs, grp)
}
\arguments{
\item{obs}{Dependent, numeric vector.}

\item{grp}{Grouping variable, i.e. \code{factor}.}
}
\description{
Games Howell post-hoc procedure
}
\references{
\url{https://rpubs.com/aaronsc32/games-howell-test}
}
\author{
\code{gitub.com/aschleg}
}
\keyword{internal}
