% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_scales.R
\name{inv}
\alias{inv}
\title{Inverting scales}
\usage{
inv(x, min, max)
}
\arguments{
\item{x}{A vector of numeric data.}

\item{min}{The minimum value of the scale.}

\item{max}{The maximum value of the scale.}
}
\description{
Inverting scales
}
\examples{
# Assuming you have a Likert-scale from 1 to 9
x <- c(4, 5, 2, 3, 7, 8, 3)
inv(x, 1, 9)
}
