% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_helpers.R
\name{labels_to_factor}
\alias{labels_to_factor}
\title{Convert all labels to factor variables}
\usage{
labels_to_factor(df)
}
\arguments{
\item{df}{A \code{data.frame}}
}
\value{
An identical \code{data.frame} with labelled data converted to factors
}
\description{
Convert all labels to factor variables
}
\examples{
\dontrun{
data \%<>\% labels_to_factor
}
}
