% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{mean_ci_sem}
\alias{mean_ci_sem}
\title{Standard Error of the Mean with CI}
\usage{
mean_ci_sem(x, conf.level = 0.95)
}
\arguments{
\item{x}{a numeric vector or R object which is coercible to one}

\item{conf.level}{the confidence level (alpha) of the Interval}
}
\value{
a \code{data.frame} with the mean, SEM and its Confidence Interval
}
\description{
Standard Error of the Mean with CI
}
\examples{
set.seed(42)
iq <- rnorm(100, 100, 15)

mean_ci_sem(iq)
}
