% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_stats.R
\name{modus}
\alias{modus}
\title{Modus}
\usage{
modus(x, na.rm = FALSE)
}
\arguments{
\item{x}{A \code{vector} with numeric data.}

\item{na.rm}{If \code{TRUE}, missing values (\code{NA}) will be ignored.}
}
\value{
A \code{vector} of length 1 of type \code{numeric} or \code{character}, depending on input.
}
\description{
Calculate the mode of a numeric vector. German name kept to avoid confusion.
}
\examples{
\dontrun{
x <- c(1, 2, 6, 2, 1, 5, 7, 8, 4, 3, 2, 2, 2)
modus(x)

# Or for nominal data
x <- structure(c(2L, 1L, 2L, 2L, 2L, 1L), .Label = c("Ja", "Nein"), class = "factor")
modus(x)
}
}

