% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ngo}
\alias{ngo}
\title{The ngo dataset}
\format{
A \code{data.frame} containing numerical and factor data.
}
\usage{
ngo
}
\description{
Sample data for teaching purposes used by Kähler (2008)
}
\note{
\code{ryouready::d.ngo} is the source of this data, but with some recoding done.
}
\references{
Kähler, W.-M. (2008). \emph{Statistische Datenanalyse: Verfahren verstehen
und mit SPSS gekonnt einsetzen}. Wiesbaden: Vieweg.
}
\keyword{dataset}
