% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom_stats.R
\name{nom_c}
\alias{nom_c}
\title{Contingency Coefficient C}
\usage{
nom_c(x, y = NULL)
}
\arguments{
\item{x}{Dependent variable. Alternatively a \code{table}.}

\item{y}{Independent variable}
}
\value{
\code{numeric} value
}
\description{
Very simple wrapper for \link[vcd]{assocstats}.
}
\examples{
nom_c(ngo$abschalt, ngo$geschl)
}
