% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom_stats.R
\name{nom_chisqu}
\alias{nom_chisqu}
\title{Simple Chi^2}
\usage{
nom_chisqu(x, y = NULL, correct = FALSE)
}
\arguments{
\item{x}{Dependent variable. Alternatively a \code{table}.}

\item{y}{Independent variable}

\item{correct}{Apply correction, passed to \code{chisq.test}.}
}
\value{
A \code{numeric} value
}
\description{
This is a very simple wrapper for \link{chisq.test}.
}
\note{
The warning message in case of low samples size and possibly incorrect
approximation is suppressed silently.
}
\examples{
nom_chisqu(ngo$abschalt, ngo$geschl)
}
