% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom_stats.R
\name{nom_lambda}
\alias{nom_lambda}
\title{Lambda}
\usage{
nom_lambda(x, y = NULL, symmetric = FALSE, reverse = FALSE)
}
\arguments{
\item{x}{Dependent variable. Alternatively a \code{table}.}

\item{y}{Independent variable}

\item{symmetric}{If \code{TRUE}, symmetric lambda is returned. Default is \code{FALSE}.}

\item{reverse}{If \code{TRUE}, row and column variable are switched.}
}
\value{
\code{numeric} value
}
\description{
Very simple wrapper for \link[DescTools:Lambda]{DescTools::Lambda}.
}
\examples{
nom_lambda(ngo$abschalt, ngo$geschl)
}
