% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom_stats.R
\name{nom_phi}
\alias{nom_phi}
\title{Phi coefficient}
\usage{
nom_phi(x, y = NULL)
}
\arguments{
\item{x}{Dependent variable. Alternatively a \code{table}.}

\item{y}{Independent variable}
}
\value{
\code{numeric} value
}
\description{
Very simple wrapper for \link[vcd]{assocstats}.
}
\examples{
nom_phi(ngo$abschalt, ngo$geschl)
}

