% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom_stats.R
\name{nom_v}
\alias{nom_v}
\title{Cramer's V}
\usage{
nom_v(x, y = NULL)
}
\arguments{
\item{x}{Dependent variable. Alternatively a \code{table}.}

\item{y}{Independent variable}
}
\value{
\code{numeric} value
}
\description{
Very simple wrapper for \link[vcd:assocstats]{vcd::assocstats}.
}
\examples{
nom_v(ngo$abschalt, ngo$geschl)
}
