% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_stats.R
\name{ord_gamma}
\alias{ord_gamma}
\title{Gamma}
\usage{
ord_gamma(x, y = NULL)
}
\arguments{
\item{x}{A \code{table} or dependent numeric variable.}

\item{y}{Empty or independent grouping variable}
}
\value{
\code{numeric} of length 1.
}
\description{
Simple wrapper for \link[ryouready:ord.gamma]{ryouready::ord.gamma}.
}
\examples{
df <- data.frame(rating = round(runif(50, 1, 5)),
                 group  = sample(c("A", "B", "C"), 50, TRUE))
tbl <- table(df)
ord_gamma(tbl)
}
