% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_stats.R
\name{ord_pairs}
\alias{ord_pairs}
\title{Retrieve all type of pairs for ordinal statistics}
\source{
Internals for this function are copied from
\href{https://gist.github.com/marcschwartz/3665743}{this gist} by GitHub user Marc Schwartz.
}
\usage{
ord_pairs(x, y = NULL)
}
\arguments{
\item{x}{Dependent variable. Alternatively a \code{table}.}

\item{y}{Independent variable}
}
\value{
A 1x5 \code{data.frame} with numeric values.
}
\description{
Retrieve all type of pairs for ordinal statistics
}
\examples{
ord_pairs(ngo$leistung, ngo$begabung)
}
