% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_stats.R
\name{ord_somers_d}
\alias{ord_somers_d}
\title{Somers' D}
\usage{
ord_somers_d(x, y = NULL, symmetric = FALSE, reverse = FALSE)
}
\arguments{
\item{x}{Dependent variable. Alternatively a \code{table}.}

\item{y}{Independent variable}

\item{symmetric}{If \code{TRUE}, symmetric D is returned. Default is \code{FALSE}.}

\item{reverse}{If \code{TRUE}, row and column variable are switched.}
}
\value{
\code{numeric} value
}
\description{
Very simple wrapper for \link[ryouready]{ord.somers.d}.
}
\examples{
ord_somers_d(ngo$abschalt, ngo$geschl)
}
