% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_stats.R
\name{ord_tau}
\alias{ord_tau}
\title{Various Tau Statistics}
\usage{
ord_tau(x, y = NULL, tau = "b", reverse = FALSE)
}
\arguments{
\item{x}{Dependent variable. Alternatively a \code{table}.}

\item{y}{Independent variable}

\item{tau}{Which of the Taus to return. Default is \code{"b"}.}

\item{reverse}{If \code{TRUE}, row and column variable are switched.}
}
\value{
\code{numeric} value
}
\description{
A wrapper for the appropriate functions form \link{DescTools} to calculate
Tau A, B and C.
}
\examples{
ord_tau(ngo$urteil, ngo$leistung, tau = "a")
}
