% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_helpers.R
\name{pval_string}
\alias{pval_string}
\title{Easy p-value formatting}
\usage{
pval_string(pv)
}
\arguments{
\item{pv}{A p-value in numeric form.}
}
\value{
A formatted \code{character} representation of the input value.
}
\description{
Easy p-value formatting
}
\note{
Simplified version of \link[pixiedust:pvalString]{pixiedust::pvalString} which considers \code{< 0.05}.
}
\examples{
pv <- c(.9, .2, .049, .009, .000003)
pval_string(pv)
}
