% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{\%<>\%}
\alias{reexports}
\alias{as_factor}
\alias{reexports}
\alias{get_label}
\alias{reexports}
\alias{get_labels}
\alias{reexports}
\alias{set_label}
\alias{reexports}
\alias{set_labels}
\alias{reexports}
\alias{word_wrap}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{haven}{\code{\link[haven]{as_factor}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%<>\%}}}

  \item{sjmisc}{\code{\link[sjmisc]{get_label}}, \code{\link[sjmisc]{get_labels}}, \code{\link[sjmisc]{set_label}}, \code{\link[sjmisc]{set_labels}}, \code{\link[sjmisc]{word_wrap}}}
}}

