% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tadaa_pairwise_t.R
\name{sidak_sd}
\alias{sidak_sd}
\title{Sidak setp-down procedure}
\usage{
sidak_sd(pvals)
}
\description{
Sidak setp-down procedure
}
\references{
\url{https://github.com/Bioconductor-mirror/multtest/blob/master/R/mt.basic.R#L81-L88}
}
\keyword{internal}
