% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tadaa_plots.R
\name{tadaa_heatmap}
\alias{tadaa_heatmap}
\title{Heatmaps}
\usage{
tadaa_heatmap(data = NULL, heat, x, group)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{heat}{The continuous variable displayed by the heat-tiles}

\item{x}{The variable on the x-axis}

\item{group}{The grouping variable (y-axis)}
}
\value{
A ggplot2 object
}
\description{
Easily generate heatmaps of a continuous variable grouped by two categorical variables.
}
\examples{
tadaa_heatmap(ngo, stunzahl, leistung, jahrgang)
}
\seealso{
Other Tadaa-plot functions: \code{\link{tadaa_int}},
  \code{\link{tadaa_mean_ci}}
}
