% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_plots.R
\name{tadaa_int}
\alias{tadaa_int}
\title{Interaction plots}
\usage{
tadaa_int(data, response, group1, group2)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{response}{Response variable}

\item{group1}{First grouping variable}

\item{group2}{Second grouping variable}
}
\value{
A list with two ggplot2 objects
}
\description{
Interaction plots
}
\examples{
tadaa_int(qmtut::ngo, stunzahl, jahrgang, geschl)
}

