% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tadaa_omnibus.R
\name{tadaa_kruskal}
\alias{tadaa_kruskal}
\title{Tadaa, Kruskal-Wallis!}
\usage{
tadaa_kruskal(formula, data = NULL, print = c("df", "console", "html",
  "markdown"))
}
\arguments{
\item{formula}{Formula for model, passed to \code{kruskal.test}.}

\item{data}{Data for model.}

\item{print}{Print method, per default a regular \code{data.frame}.
Otherwise passed to \link[pixiedust:sprinkle_print_method]{pixiedust::sprinkle_print_method} for fancyness.}
}
\value{
A \code{data.frame} by default, otherwise \code{dust} object, depending on \code{print}.
}
\description{
Tadaa, Kruskal-Wallis!
}
\examples{
tadaa_kruskal(stunzahl ~ jahrgang, data = ngo)
}
\seealso{
Other Tadaa-functions: \code{\link{tadaa_aov}},
  \code{\link{tadaa_chisq}}, \code{\link{tadaa_levene}},
  \code{\link{tadaa_nom}}, \code{\link{tadaa_normtest}},
  \code{\link{tadaa_one_sample}}, \code{\link{tadaa_ord}},
  \code{\link{tadaa_pairwise_gh}},
  \code{\link{tadaa_pairwise_tukey}},
  \code{\link{tadaa_pairwise_t}},
  \code{\link{tadaa_t.test}}, \code{\link{tadaa_wilcoxon}}
}
