% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom_stats.R
\name{tadaa_nom}
\alias{tadaa_nom}
\title{Get all the nominal stats}
\usage{
tadaa_nom(x, y = NULL, round = 2, print = "console")
}
\arguments{
\item{x}{Dependent variable. Alternatively a \code{table}.}

\item{y}{Independent variable}

\item{round}{Ho many digits should be rounded. Default is 2.}

\item{print}{Print method. Passed to \link[pixiedust]{sprinkle_print_method} as of now.}
}
\value{
A \code{dust} object, depending on \code{print}.
}
\description{
Get all the nominal stats
}
\examples{
tadaa_nom(ngo$abschalt, ngo$geschl)
}

