% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tadaa_pairwise.R
\name{tadaa_pairwise_t}
\alias{tadaa_pairwise_t}
\title{Extended Pairwise t-Tests}
\usage{
tadaa_pairwise_t(data, response, group1, group2 = NULL, p.adjust = "bonf",
  paired = FALSE, pool.sd = !paired, alternative = "two.sided",
  print = "df")
}
\arguments{
\item{data}{A \code{data.frame} containing the variables.}

\item{response}{The response variable, i.e. the dependent numeric vector.}

\item{group1}{The grouping variables, typically a \code{factor}.}

\item{group2}{(Optional) second grouping variable.}

\item{p.adjust}{The p-adjustment method, see \link[stats]{p.adjust.methods}, passed
to \link[stats]{pairwise.t.test}.}

\item{paired}{Defaults to \code{FALSE}, also passed to \link[stats]{pairwise.t.test}.}

\item{pool.sd}{Defaults to the inverse of \code{paired}, passed
to \link[stats]{pairwise.t.test}.}

\item{alternative}{Defaults to \code{two.sided}, also passed to \link[stats]{pairwise.t.test}.}

\item{print}{Print method, defaults to \code{df} for \code{data.frame} output, otherwise
passed to \link[pixiedust]{sprinkle_print_method}.}
}
\value{
A \code{data.frame} with columns \code{term}, \code{comparison} and \code{adj.p.value}.
}
\description{
This is an extension of \link[stats]{pairwise.t.test} that's meant to deal with
interactions our of the box, while also performing pairwise tests for the primary terms.
The output of the function is modeled after \link[stats]{TukeyHSD}, unfortunately
without confidence intervals or test statistic though.
}
\note{
The adjustment method is applied within each \code{term}, meaning that the number of
pairwise t-tests counted for the adjustment is only equal to the number of rows
per \code{term} of the output.
}
\examples{
tadaa_pairwise_t(ngo, deutsch, jahrgang, geschl, print = "console")

}
\seealso{
Other Tadaa-functions: \code{\link{tadaa_aov}},
  \code{\link{tadaa_kruskal}}, \code{\link{tadaa_nom}},
  \code{\link{tadaa_normtest}},
  \code{\link{tadaa_one_sample}}, \code{\link{tadaa_ord}},
  \code{\link{tadaa_t.test}}, \code{\link{tadaa_wilcoxon}}
}
