% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tadaatoolbox-package.r
\docType{package}
\name{tadaatoolbox-package}
\alias{tadaatoolbox}
\alias{tadaatoolbox-package}
\title{tadaatoolbox: Helpers for Data Analysis and Presentation Focused on Undergrad Psychology}
\description{
Contains functions for the easy display of statistical tests as well as
    some convenience functions for data cleanup. It is meant to ease existing workflows
    with packages like 'sjPlot', 'dplyr', and 'ggplot2'. The primary components are the functions
    prefixed with 'tadaa_', which are built to work in an interactive environment, but also print
    tidy markdown tables powered by 'pixiedust' for the creation of 'RMarkdown' reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tadaadata/tadaatoolbox}
  \item Report bugs at \url{https://github.com/tadaadata/tadaatoolbox/issues}
}

}
\author{
\strong{Maintainer}: Lukas Burk \email{lukas@tadaa-data.de} (0000-0001-7528-3795)

Authors:
\itemize{
  \item Tobias Anton \email{tobi@tadaa-data.de}
}

Other contributors:
\itemize{
  \item Daniel Lüdecke \email{d.luedecke@uke.de} [contributor]
  \item Gesa Graf \email{ggraf@uni-bremen.de} [contributor]
}

}
\keyword{internal}
