% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.R
\name{theme_latex}
\alias{theme_latex}
\title{ggplot2 theme using the Computer Modern font}
\usage{
theme_latex()
}
\arguments{
\item{base_size}{Base text size.}

\item{base_family}{Base text family.}
}
\value{
A ggplot2 theme
}
\description{
ggplot2 theme using the Computer Modern font
}
\examples{
ggplot2::qplot(1:10, 1:10, geom = "point") + theme_latex()
}

