% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.R
\name{theme_readthedown}
\alias{theme_readthedown}
\title{ggplot2 theme to fit the readthedown Rmd format}
\usage{
theme_readthedown(base_size = 12, base_family = "", bg = "#fcfcfc",
  axis_emph = "xy")
}
\arguments{
\item{base_size}{Base text size.}

\item{base_family}{Base text family. Use \code{"Roboto Slab"} to match the \code{readthedown}
headers.}

\item{bg}{Background color, defaults to \code{\link[rmdformats]{readthedown}}'s background,
\code{#fcfcfc}}

\item{axis_emph}{Which axis to emphasize visually (black color).}
}
\value{
A ggplot2 theme
}
\description{
A ggplot theme to fit \code{\link[rmdformats]{readthedown}} in terms of background color and
dark grid lines.
}
\examples{
ggplot2::qplot(1:10, 1:10, geom = "point") + theme_readthedown()

}

