% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_test.R
\name{z_test}
\alias{z_test}
\title{One- and Two-Sample z-Test}
\usage{
z_test(x, y = NULL, alternative = c("two.sided", "less", "greater"),
  mu = 0, sigma_x, sigma_y, paired = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{A (non-empty) numeric vector of data values}

\item{y}{An optional (non-empty) numeric vector of data values.}

\item{alternative}{A character string specifying the alternative hypothesis, must be
one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{mu}{A logical indicating whether you want a paired z-test.}

\item{paired}{A logical indicating whether you want a paired t-test.}

\item{conf.level}{Confidence level of the interval.}

\item{...}{}
}
\value{
An object of class `htest``
}
\description{
One- and Two-Sample z-Test
}
\examples{
x <- 1:10
y <- 1:10 + rnorm(10, 3)
z_test(x, y)
}
