% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{tadaa_aov}
\alias{tadaa_aov}
\title{Tadaa, anova!}
\usage{
tadaa_aov(formula, data = NULL, show_effect_size = TRUE,
  print = "console")
}
\arguments{
\item{formula}{Formula for model, passed to \code{aov}.}

\item{data}{Data for model.}

\item{show_effect_size}{If \code{TRUE} (default), effect size is appended as a column.}

\item{print}{Print method. Passed to \link[pixiedust]{sprinkle_print_method} as of now.}
}
\value{
A \code{dust} object, depending on \code{print}.
}
\description{
Tadaa, anova!
}
\examples{
tadaa_aov(stunzahl ~ jahrgang, data = ngo)
}

