% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_plots.R
\name{tadaa_int}
\alias{tadaa_int}
\title{Interaction plots}
\usage{
tadaa_int(data, response, group1, group2)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{response}{Response variable}

\item{group1}{First grouping variable}

\item{group2}{Second grouping variable}
}
\value{
Invisible: A list with two ggplot2 objects named \code{p1} and \code{p2}.
Printed: The two ggplot2 objects.
}
\description{
Easily generate interaction plots of two nominal grouping
variables and a numeric response variable.
}
\examples{
tadaa_int(ngo, stunzahl, jahrgang, geschl)
}

