% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_plots.R
\name{tadaa_int}
\alias{tadaa_int}
\title{Interaction plots}
\usage{
tadaa_int(data, response, group1, group2, brewer_palette = "Set1")
}
\arguments{
\item{data}{A \code{data.frame}}

\item{response}{Response variable}

\item{group1}{First grouping variable}

\item{group2}{Second grouping variable}

\item{brewer_palette}{The name of the \link[RColorBrewer]{RColorBrewer} palette to use, defaults to \code{Set1}}
}
\value{
Invisible: A list with two ggplot2 objects named \code{p1} and \code{p2}.
Printed: The two ggplot2 objects.
}
\description{
Easily generate interaction plots of two nominal grouping
variables and a numeric response variable.
}
\examples{
tadaa_int(ngo, stunzahl, jahrgang, geschl)
}
\seealso{
Other Tadaa.functions: \code{\link{tadaa_aov}},
  \code{\link{tadaa_nom}}, \code{\link{tadaa_ord}},
  \code{\link{tadaa_t.test}}
}

