% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tadaa_plots.R
\name{tadaa_mean_ci}
\alias{tadaa_mean_ci}
\title{Means with Errorbars}
\usage{
tadaa_mean_ci(data, response, group, brewer_palette = NULL)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{response}{Response variable, numeric.}

\item{group}{Grouping variable, ideally a \code{factor}.}

\item{brewer_palette}{Optional: The name of the \link[RColorBrewer]{RColorBrewer} palette to use,
defaults to \code{NULL}: No brewer palette.}
}
\value{
A ggplot2 object.
}
\description{
Means with Errorbars
}
\examples{
tadaa_mean_ci(ngo, deutsch, jahrgang, brewer_palette = "Set1")
}
\seealso{
Other Tadaa-plot functions: \code{\link{tadaa_heatmap}},
  \code{\link{tadaa_int}}
}
