% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tadaa_plots.R
\name{tadaa_mean_ci}
\alias{tadaa_mean_ci}
\title{Plot Means with Errorbars}
\usage{
tadaa_mean_ci(data, response, group, brewer_palette = "Set1")
}
\arguments{
\item{data}{A \code{data.frame}}

\item{response}{Response variable, numeric.}

\item{group}{Grouping variable, ideally a \code{factor}.}

\item{brewer_palette}{Optional: The name of the \code{RColorBrewer} palette to use,
defaults to \code{Set1}. Use \code{NULL} for no brewer palette.}
}
\value{
A ggplot2 object.
}
\description{
Plot Means with Errorbars
}
\examples{
tadaa_mean_ci(ngo, deutsch, jahrgang, brewer_palette = "Set1")
}
\seealso{
Other Tadaa-plot functions: \code{\link{tadaa_balance}},
  \code{\link{tadaa_int}}, \code{\link{tadaa_plot_tukey}}
}
