% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_stats.R
\name{tadaa_ord}
\alias{tadaa_ord}
\title{Get all the ordinal stats}
\usage{
tadaa_ord(x, y = NULL, round = 2, print = "console")
}
\arguments{
\item{x}{Dependent variable. Alternatively a \code{table}.}

\item{y}{Independent variable}

\item{round}{Ho many digits should be rounded. Default is 2.}

\item{print}{Print method. Passed to \link[pixiedust]{sprinkle_print_method} as of now.}
}
\value{
A \code{dust} object, depending on \code{print}.
}
\description{
As of now, only Gamma and Somers D are supported. But let's be honest: Everybody hates Tau.
}
\examples{
tadaa_ord(ngo$abschalt, ngo$geschl)
}
\seealso{
Other Tadaa.functions: \code{\link{tadaa_aov}},
  \code{\link{tadaa_int}}, \code{\link{tadaa_nom}},
  \code{\link{tadaa_t.test}}
}

