% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tadaa_stats.R
\name{tadaa_t.test}
\alias{tadaa_t.test}
\title{Tadaa, t-Test!}
\usage{
tadaa_t.test(data, response, group, direction = "two.sided", paired = FALSE,
  print = "df")
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{response}{The response variable (dependent).}

\item{group}{The group variable, usually a \code{factor}.}

\item{direction}{Test direction, like \code{alternative} in \link{t.test}.}

\item{paired}{If \code{TRUE}, a paired t.test is performed.}

\item{print}{Print method, per default a regular \code{data.frame}.
Otherwise passed to \link[pixiedust]{sprinkle_print_method} for fancyness.}
}
\value{
A \code{data.frame} by default, otherwise \code{dust} object,
depending on \code{print}.
}
\description{
An extension for \link[stats]{t.test} with added boni and tidy and/or pretty output.
Before a t-test is performed, \link[car]{leveneTest} is consulted as to wether
heteroskedasticity is present (using the default \code{center = "mean"} method for
a more robust test), and sets \code{var.equal} accordingly.
Afterwards, the effect size is calculated and \link[pwr]{pwr.t.test} or
\link[pwr]{pwr.t2n.test} are used to calculate the test's power accordingly.
The result is either returned as a \link[broom]{tidy} \code{data.frame} or prettified using
various \link[pixiedust]{sprinkle} shenanigans.
}
\note{
The cutoff for the interal Levene's test to decided whether or not to perform
a Welch-corrected t-test is set to `0.1` by default.
}
\examples{
df <- data.frame(x = runif(100), y = sample(c("A", "B"), 100, TRUE))
tadaa_t.test(df, x, y)

df <- data.frame(x = runif(100), y = c(rep("A", 50), rep("B", 50)))
tadaa_t.test(df, x, y, paired = TRUE)

tadaa_t.test(ngo, deutsch, geschl, print = "console")
}
\seealso{
Other Tadaa-functions: \code{\link{tadaa_aov}},
  \code{\link{tadaa_kruskal}}, \code{\link{tadaa_nom}},
  \code{\link{tadaa_normtest}},
  \code{\link{tadaa_one_sample}}, \code{\link{tadaa_ord}},
  \code{\link{tadaa_pairwise_gh}},
  \code{\link{tadaa_pairwise_t}},
  \code{\link{tadaa_wilcoxon}}
}
